// src/admin/shared/interfaces/object-linked-type.enum.ts
export enum ObjectLinkedType {
  article = 'article',
  celebrity = 'celebrity',
  checklist = 'checklist',
  coaching = 'coaching',
  course = 'course',
  courseStep = 'courseStep',
  exercise = 'exercise',
  habit = 'habit',
  habitCategory = 'habitCategory',
  interestingCategory = 'interestingCategory',
  letter = 'letter',
  task = 'task',
}